Merge APIs
The cube-api has been merged into the learning-object-submission API. This is to create a single API Gateway for the full platform, counter to the Backend for Front-end pattern.

The two APIs have been merged, but are still segmented. Requests for cube must prepend '/cube' to the route. Extensive refactoring is needed, in which the cube routes should be merged into the base layer of the API.

Contributes to #6